Moved files to appropriate directories
