+++ app/src/main/java/com/protify/protifyapp/features/login/RegisterActivity.kt
@@ -1,4 +1,4 @@
-package com.protify.protifyapp
+package com.protify.protifyapp.features.login
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Column
